Complete Refactor of Plugin Structure / Move StickyHeader to own class
