+++ esbuild.config.mjs
@@ -43,7 +43,7 @@ const context = await esbuild.context({
   outfile: 'main.js',
   plugins: [
     esbuildSvelte({
-      compilerOptions: { css: true },
+      compilerOptions: { css: 'injected' },
       preprocess: sveltePreprocess(),
     }),
   ],
