--- src/defaultSetting.ts
@@ -1,4 +1,0 @@
-export default {
-  max: 0,
-  mode: 'default',
-} satisfies ISetting;
