+++ src/typing.d.ts
@@ -1,4 +1,18 @@
-interface ISetting {
+import 'obsidian';
+
+export interface ISetting {
   max: number;
   mode: 'default' | 'concise';
+}
+
+declare module 'obsidian' {
+  interface WorkspaceLeaf {
+    id: string;
+  }
+  interface MarkdownView {
+    getFile: () => TFile;
+    editMode: {
+      sourceMode: boolean;
+    };
+  }
 }
