+++ src/ui/store.ts
@@ -1,3 +1,5 @@
 import { writable } from 'svelte/store';
 
 export const headings = writable<string[]>([]);
+
+export const editMode = writable<boolean>(false);
